apps/api/pyproject.toml
[project]
name = "parsegrid-api"
version = "0.1.0"
description = "ParseGrid — AI-Powered Unstructured Data Extraction Pipeline API"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    # Validation & Config
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    # Auth
    "pyjwt>=2.9.0",
    # Task queue
    "celery[redis]>=5.4.0",
    "redis>=5.0.0",
    # S3-compatible storage
    "boto3>=1.35.0",
    # AI / LLM
    "openai>=1.60.0",
    "openai-agents>=0.1.0",
    # OCR (local, air-gapped)
    "paddlepaddle>=3.0.0",
    "paddleocr>=2.9.0",
    # PDF processing
    "pymupdf>=1.25.0",
    # Utilities
    "python-multipart>=0.0.18",
    "httpx>=0.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["app*"]
